readability-redundant-casting
warning: redundant explicit casting to the same type as the sub-expression, remove this casting

+++ CvGameCoreDLL_Expansion2/CvBuildingClasses.cpp
@@ -6643,7 +6643,7 @@ void BuildingArrayHelpers::Write(FDataStream& kStream, int* paiBuildingArray, in
 		}
 		else
 		{
-			kStream << (int)0;
+			kStream << 0;
 		}
 	}
 }
